hellodjango/lipaiapi/requirements.txt
requests
allure-pytest
pytest
pytest-ordering
ddddocr

